add variables to support local init rpm (#40)
* add variables to support local init rpm and pgp key override
* fmt
* remove pgp key override, let install-additional-packages.sh handle local init install
* remove standalone key file

+++ scripts/install-ecs-init.sh
@@ -6,6 +6,11 @@ if [ -n "$AIR_GAPPED" ]; then
     exit 0
 fi
 
+if [ -n "$INIT_LOCAL_OVERRIDE" ]; then
+    echo "ecs-init is provided locally, assuming it's in additional-packages/ directory"
+    exit 0
+fi
+
 WORK_DIR="$(mktemp -d)"
 trap "rm -rf ${WORK_DIR}" EXIT
